chore: extract logic
